Description of Project
This was the project that I worked on in 'Tools for Games and Animation' which is a small text based tactical game where you control three different units with different classes which are 'knights, tanks and mages' and then you send out any of them depending on the assigned input. You will battle multiple enemies which will be the same type of class as your units until you defeated all of them. You decide on who is sent out based on the preferred match up as each class will deal more damage against a particular enemy, for example mages will deal more damage to tanks.There are four different options to choose, three being assigned to each of your units and then one choice is to defend against the enemy attack. 
This game uses classes to define what each of the units are and their stats using a constructor to initialise each of the values. This allows for them to be called easier in the main body of the program when the game is running. For example, by typing 'tank1.Tankhp', the program will recognise what class you are referring to and the values to go along with it. 
Other technique that this project uses are while loops and switch case statements. The while loops allows for the battles to keep on going by checking for a condition to see if the player's hp and the enemy's hp is above 0. If the enemy hp is 0 then the player will move on to the next battle, if the player hp is 0 then it is game over. This will allow the battles to keep on going until either of the conditions have been met. The technique that I used is switch case statement which allows me to set up the different outcomes for the player depending on what choice they chose.